Add -Wconversion to clang; fix errors found
- Make all enums based on size_t
- Format enums in a regular way (one value per line, etc.)
- Make signedness conversions explicit
- static_cast<int> becomes static_cast<size_t> when result is an array
  index

+++ include/Utility/Fixed_Capacity_Vector.h
@@ -57,7 +57,7 @@ class Fixed_Capacity_Vector
         template<typename Iterator>
         constexpr auto scoped_push_back(const Iterator begin, const Iterator end) noexcept
         {
-            assert(size() + std::distance(begin, end) <= data.size());
+            assert(size() + size_t(std::distance(begin, end)) <= data.size());
             return Scoped_Push_Guard(*this, begin, end);
         }
